Break everything, refactor. Longer description below...
Since we based this off the react-redux-starter-kit and decided we're taking
a different direction, this (rather large) commit is the first of many to
clean up the webpack layer. We're not going for complete configurability
but rather for a simple and understandable set of configurations (for
development at the moment but also for production eventually).

We're also planning on simplifying the directory structure.