Intended Scope of Work: List all npm modules, grouped by dependency type, in ascending order and provide short description - [Status Completed]
As per @mcalthrop enhancement #6  request

>Originally posted by @mcalthrop on Jan 11, 2016, 11:27 AM MST

```
There are a lot of modules used in this boilerplate code.

To help a developer understand why various modules have been chosen, it would be useful to have a list of them, and a brief description of why each is used.

This should be simple enough to implement as a bulleted list in the README; may end up as a separate .md file.
```